+++ doc/src/changelog.md
@@ -9,6 +9,10 @@
   next argument to the specified command. For example, using `#!python3 -c` as the
   first line in a custom command will cause `python3 -c <command>` to be executed.
 
+- Trees can now use branches defined in separate remotes when configuring the
+  default branch to checkout. `garden grow` will now fetch the remote associated with the
+  configured branch switching branches in order to make this possible.
+
 **Development**:
 
 - `garden` can now be built on Windows. Symlink trees and the XDG base directory support